+++ src/Service/ComuniService.php
@@ -33,6 +33,29 @@ class ComuniService
     }
 
     /**
+     * 
+     * Retrieves a list of distinct comune names (denominazioneIta) from the MetareteComune entity associated with the specified province.
+     *
+     * This method constructs a query using the entity manager to select distinct comune names associated with the specified province,
+     * orders them in ascending order, and returns the result as an array of strings.
+     *
+     * @return array<string> An array of distinct comune names (denominazioneIta).
+     */
+    public function getComuniListFromProvince(string $provincia): array
+    {
+        $query = $this->entityManager->createQueryBuilder()
+            ->select('DISTINCT c.denominazioneIta')
+            ->from('Metarete\ComuniBundle\Entity\MetareteComune', 'c')
+            ->where('c.siglaProvincia = :provincia')
+            ->setParameter('provincia', $provincia)
+            ->orderBy('c.denominazioneIta', 'ASC')
+            ->getQuery();
+        $result = $query->getResult();
+
+        return array_map(fn($item): string => $item['denominazioneIta'], $result);
+    }
+
+    /**
      * Retrieves a list of distinct province abbreviations (siglaProvincia) from the MetareteComune entity.
      *
      * This method constructs a query using the entity manager to select distinct province abbreviations,
@@ -45,6 +68,53 @@ class ComuniService
         $query = $this->entityManager->createQueryBuilder()
             ->select('DISTINCT c.siglaProvincia')
             ->from('Metarete\ComuniBundle\Entity\MetareteComune', 'c')
+            ->orderBy('c.siglaProvincia', 'ASC')
+            ->getQuery();
+        $result = $query->getResult();
+
+        return array_map(fn($item): string => $item['siglaProvincia'], $result);
+    }
+
+    /**
+     * 
+     * Retrieves a province abbreviation (siglaProvincia) from the MetareteComune entity associated to the given Comune.
+     *
+     * This method constructs a query using the entity manager to select a province abbreviation associated to the given Comune,
+     * orders them in ascending order, and returns the result as an array of strings.
+     *
+     * @param string $comune The codice istat or name of the comune.
+     * @return array<string> An array of distinct province abbreviations (siglaProvincia).
+     */
+    public function getProvinceFromComune(string $comune): array
+    {
+        $query = $this->entityManager->createQueryBuilder()
+            ->select('DISTINCT c.siglaProvincia')
+            ->from('Metarete\ComuniBundle\Entity\MetareteComune', 'c')
+            ->where('c.codiceIstat = :comune OR c.denominazioneIta = :comune')
+            ->setParameter('comune', $comune)
+            ->orderBy('c.siglaProvincia', 'ASC')
+            ->getQuery();
+        $result = $query->getResult();
+
+        return array_map(fn($item): string => $item['siglaProvincia'], $result);
+    }
+
+    /**
+     * 
+     * Retrieves a province abbreviation (siglaProvincia) from the MetareteComune entity associated to the given CAP.
+     *
+     * This method constructs a query using the entity manager to select a province abbreviation associated to the given CAP,
+     * orders them in ascending order, and returns the result as an array of strings.
+     *
+     * @return array<string> An array of distinct province abbreviations (siglaProvincia).
+     */
+    public function getProvinceFromCAP(string $cap): array
+    {
+        $query = $this->entityManager->createQueryBuilder()
+            ->select('DISTINCT c.siglaProvincia')
+            ->from('Metarete\ComuniBundle\Entity\MetareteComune', 'c')
+            ->where('c.cap = :cap')
+            ->setParameter('cap', $cap)
             ->orderBy('c.siglaProvincia', 'ASC')
             ->getQuery();
         $result = $query->getResult();
@@ -100,6 +170,9 @@ class ComuniService
     }
 
     /**
+     * 
+     * @deprecated version 1.0.5
+     * 
      * Retrieves the codice ISTAT of a comune given its exact name and CAP.
      *
      * This method queries the database to get the codice istat of a comune given its name and CAP.
@@ -117,6 +190,51 @@ class ComuniService
             ->where('c.denominazioneIta = :comune')
             ->andWhere('c.cap = :cap')
             ->setParameter('comune', $comune)
+            ->setParameter('cap', $cap)
+            ->getQuery();
+        $result = $query->getOneOrNullResult();
+
+        return ($result ? $result['codiceIstat'] : null);
+    }
+
+    /*
+     * Retrieves the codice ISTAT of a comune given its exact name
+     *
+     * This method queries the database to get the codice istat of a comune given its name.
+     * The result is a string containing the codice istat or null if no match is found.
+     *
+     * @param string $comune The name of the comune.
+     * @return string|null The codice istat of the comune or null if no match is found.
+     */
+    public function getCodiceISTATFromComune(string $comune): string|null
+    {
+        $query = $this->entityManager->createQueryBuilder()
+            ->select('DISTINCT c.codiceIstat')
+            ->from('Metarete\ComuniBundle\Entity\MetareteComune', 'c')
+            ->where('c.denominazioneIta = :comune')
+            ->setParameter('comune', $comune)
+            ->getQuery();
+        $result = $query->getOneOrNullResult();
+
+        return ($result ? $result['codiceIstat'] : null);
+    }
+
+    /*
+     * Retrieves the codice ISTAT of a comune given its CAP.
+     *
+     * This method queries the database to get the codice istat of a comune given its CAP.
+     * The result is a string containing the codice istat or null if no match is found.
+     *
+     * @param string $comune The name of the comune.
+     * @param string $cap The CAP of the comune.
+     * @return string|null The codice istat of the comune or null if no match is found.
+     */
+    public function getCodiceISTATFromCap(string $cap): string|null
+    {
+        $query = $this->entityManager->createQueryBuilder()
+            ->select('DISTINCT c.codiceIstat')
+            ->from('Metarete\ComuniBundle\Entity\MetareteComune', 'c')
+            ->where('c.cap = :cap')
             ->setParameter('cap', $cap)
             ->getQuery();
         $result = $query->getOneOrNullResult();